requirements/dev.txt
-r main.txt
pprint
sphinx
sphinx-rtd-theme
m2r
pytest
blessed
pylint
parameterized
vcrpy
